+++ src/library/renderhud/RenderHUD_SDL2_renderer.cpp
@@ -21,7 +21,6 @@
 
 #include "logging.h"
 #include "hook.h"
-#include "screencapture/ScreenCapture.h"
 #include "GlobalState.h"
 
 #include "../external/imgui/imgui.h"