+++ src/library/renderhud/RenderHUD_Vulkan.cpp
@@ -21,7 +21,6 @@
 
 #include "logging.h"
 #include "hook.h"
-#include "screencapture/ScreenCapture.h"
 #include "GlobalState.h"
 #include "rendering/vulkanwrappers.h"